creatd a main script to run all functions, and created tests for the new functions
